ember-drf-example/requirements.txt
Django==1.7.1
djangorestframework==3.0.1
django_filter==0.9.1
